refactor: unify graphing apps to share more code
