+++ packages/swingset-runner/bin/graphMem
@@ -1,8 +1,1 @@
-#!/usr/bin/env -S node -r esm
-
-/**
- * Simple boilerplate program providing linkage to launch an application written using modules within the
- * as yet not-entirely-ESM-supporting version of NodeJS.
- */
-import { main } from '../src/graphMem.js';
-main();
+../src/graphMem.js
